Dividido herança de objetos entre os arquivos l10n_br_account/account_invoice.py l10n_br_account_product/account_invoice.py, l10n_br_account_service/account_invoice.py e l10n_br_account_product_service/account_invoice.py
